requirements.txt
pandas==2.1.4
requests==2.31.0
tqdm==4.66.1
scikit-image==0.22.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
scikit-learn==1.3.2
tslearn==0.6.3